+++ BTrees/tests/common.py
@@ -2354,6 +2354,7 @@ class SetResult(object):
                 C = self.union(A, B)
                 self.assertTrue(not hasattr(C, "values"))
                 self.assertEqual(list(C), self._union(A, B))
+                self.assertEqual(set(A) | set(B), set(A | B))
 
     def testIntersection(self):
         inputs = self.As + self.Bs
@@ -2362,6 +2363,7 @@ class SetResult(object):
                 C = self.intersection(A, B)
                 self.assertTrue(not hasattr(C, "values"))
                 self.assertEqual(list(C), self._intersection(A, B))
+                self.assertEqual(set(A) & set(B), set(A & B))
 
     def testDifference(self):
         inputs = self.As + self.Bs
@@ -2375,6 +2377,7 @@ class SetResult(object):
                     self.assertEqual(list(C.items()), want)
                 else:
                     self.assertEqual(list(C), want)
+                self.assertEqual(set(A) - set(B), set(A - B))
 
     def testLargerInputs(self): # pylint:disable=too-many-locals
         from BTrees.IIBTree import IISet # pylint:disable=no-name-in-module